src: run alongside fbdev
The old code checked for whether or not files matching
/dev/fb* or /dev/dri/card* were present on a system.

This is wrong for multiple reasons.
Vesa works fine alongside a framebuffer driver.
Testing whether a device file exists or not is not
enough to detarmine that a driver is present, those
device files should also be able to be opened.

Signed-off-by: stefan11111 <stefan11111@shitposting.expert>